@@ -1,113 +1,96 @@
 {
   "name": "ulearn",
-  "version": "0.1.0",
+  "version": "0.2.0",
   "private": true,
   "scripts": {
-    "start": "node scripts/start.js",
-    "build": "node scripts/build.js",
-    "analyze": "node scripts/build.js analyze",
+    "start": "webpack serve -c config/webpack.config.dev.ts --open --port 3000",
+    "build": "webpack -c config/webpack.config.prod.ts --node-env production",
+    "analyze": "webpack -c config/webpack.config.prod.analyze.ts --node-env production",
     "test": "jest",
-    "storybook": "start-storybook -p 9001 -c .storybook -s public"
+    "storybook": "start-storybook -p 9001 -c .storybook -s public",
+    "outdated": "yarn outdated"
   },
   "dependencies": {
-    "@babel/core": "7.16.0",
-    "@babel/eslint-parser": "7.16.3",
-    "@babel/plugin-proposal-class-properties": "7.16.0",
-    "@babel/plugin-proposal-object-rest-spread": "7.16.0",
-    "@babel/plugin-proposal-private-methods": "7.16.0",
-    "@babel/plugin-transform-arrow-functions": "7.16.0",
-    "@babel/plugin-transform-modules-commonjs": "7.16.0",
-    "@babel/preset-env": "7.16.4",
-    "@babel/preset-react": "7.16.0",
-    "@babel/preset-typescript": "7.16.0",
-    "@microsoft/signalr": "5.0.12",
-    "@sentry/browser": "6.16.0",
-    "@sentry/react": "6.16.0",
-    "@sentry/tracing": "6.16.0",
-    "@skbkontur/react-icons": "4.4.0",
-    "@skbkontur/react-ui": "3.9.3",
-    "@skbkontur/react-ui-validations": "1.5.7",
-    "autoprefixer": "10.4.0",
-    "babel-eslint": "10.1.0",
-    "babel-loader": "8.2.3",
-    "babel-preset-react-app": "10.0.0",
+    "@babel/core": "7.18.5",
+    "@babel/eslint-parser": "7.18.2",
+    "@babel/preset-env": "7.18.2",
+    "@babel/preset-react": "7.17.12",
+    "@babel/preset-typescript": "7.17.12",
+    "@microsoft/signalr": "6.0.6",
+    "@sentry/browser": "7.2.0",
+    "@sentry/react": "7.2.0",
+    "@sentry/tracing": "7.2.0",
+    "@skbkontur/react-icons": "4.5.0",
+    "@skbkontur/react-ui": "4.1.1",
+    "@skbkontur/react-ui-validations": "1.8.4",
+    "autoprefixer": "10.4.7",
+    "babel-loader": "8.2.5",
+    "babel-preset-react-app": "10.0.1",
     "bootstrap": "3.3.7",
-    "bootstrap-fileinput": "5.2.6",
+    "bootstrap-fileinput": "5.2.8",
     "bootstrap-select": "1.13.18",
     "case-sensitive-paths-webpack-plugin": "2.4.0",
-    "chalk": "4.1.2",
     "classnames": "2.3.1",
-    "clipboard": "2.0.8",
-    "codemirror": "5.64.0",
-    "core-js": "3.19.3",
-    "css-loader": "6.5.1",
-    "diff": "5.0.0",
+    "codemirror": "5.65.6",
+    "core-js": "3.23.2",
+    "css-loader": "6.7.1",
     "diff-match-patch": "1.0.5",
-    "dotenv": "10.0.0",
-    "dotenv-expand": "5.1.0",
     "file-loader": "6.2.0",
     "file-saver": "2.0.5",
-    "flux": "4.0.2",
-    "fs-extra": "10.0.0",
-    "highcharts": "9.3.2",
+    "flux": "4.0.3",
+    "fs-extra": "10.1.0",
+    "highcharts": "10.1.0",
     "html-webpack-plugin": "5.5.0",
     "identity-obj-proxy": "3.0.0",
     "js-cookie": "3.0.1",
-    "katex": "0.15.1",
-    "less": "4.1.2",
-    "less-loader": "10.2.0",
-    "likely": "0.2.0",
-    "mini-css-extract-plugin": "2.4.5",
-    "moment": "2.29.1",
+    "katex": "0.16.0",
+    "less": "4.1.3",
+    "less-loader": "11.0.0",
+    "mini-css-extract-plugin": "2.6.1",
+    "moment": "2.29.3",
     "moment-timezone": "0.5.34",
     "outdated-browser-rework": "3.0.1",
-    "postcss": "8.4.4",
-    "postcss-flexbugs-fixes": "5.0.2",
-    "postcss-loader": "6.2.1",
-    "postcss-preset-env": "6.7.0",
-    "prettier": "2.5.1",
-    "prop-types": "15.7.2",
-    "query-string": "7.0.1",
-    "raf": "3.4.1",
-    "react": "17.0.2",
-    "react-base16-styling": "0.8.1",
+    "postcss": "8.4.14",
+    "postcss-loader": "7.0.0",
+    "postcss-preset-env": "7.7.1",
+    "prop-types": "15.8.1",
+    "query-string": "7.1.1",
+    "react": "18.2.0",
+    "react-base16-styling": "0.9.1",
     "react-codemirror2": "7.2.1",
     "react-cookie": "4.1.1",
-    "react-copy-to-clipboard": "5.0.4",
-    "react-dev-utils": "12.0.0-next.58",
-    "react-dom": "17.0.2",
+    "react-dev-utils": "12.0.1",
+    "react-dom": "18.2.0",
     "react-helmet": "6.1.0",
-    "react-image-gallery": "1.2.7",
+    "react-image-gallery": "1.2.8",
     "react-lifecycles-compat": "3.0.4",
-    "react-redux": "7.2.6",
+    "react-redux": "8.0.2",
     "react-responsive": "8.2.0",
-    "react-router-dom": "5.3.0",
-    "react-textarea-autosize": "8.3.3",
+    "react-router-dom": "6.3.0",
+    "react-textarea-autosize": "8.3.4",
     "react-transition-group": "4.4.2",
     "react-yandex-metrika": "2.6.0",
-    "react-youtube": "7.13.1",
-    "redux": "4.1.2",
+    "react-youtube": "9.0.2",
+    "redux": "4.2.0",
     "redux-thunk": "2.4.1",
     "style-loader": "3.3.1",
-    "typescript": "4.5.2",
+    "typescript": "4.7.4",
     "url-loader": "4.1.1",
-    "webpack": "5.65.0",
+    "webpack": "5.73.0",
     "webpack-bundle-analyzer": "4.5.0",
-    "webpack-cli": "4.9.1",
-    "webpack-dev-server": "4.6.0",
     "webpack-jquery-ui": "2.0.1",
-    "webpack-manifest-plugin": "4.0.2",
+    "webpack-manifest-plugin": "5.0.0",
     "webpack-merge": "5.8.0",
     "webpack-pwa-manifest": "4.3.0",
-    "workbox-webpack-plugin": "6.4.2"
+    "workbox-webpack-plugin": "6.5.3"
   },
   "devDependencies": {
-    "@storybook/addon-essentials": "6.4.9",
-    "@storybook/addon-viewport": "6.4.9",
-    "@storybook/addons": "6.4.9",
-    "@storybook/builder-webpack5": "6.4.9",
-    "@storybook/manager-webpack5": "6.4.0",
-    "@storybook/react": "6.4.9",
+    "@storybook/addon-essentials": "6.5.9",
+    "@storybook/addon-viewport": "6.5.9",
+    "@storybook/addons": "6.5.9",
+    "@storybook/builder-webpack5": "6.5.9",
+    "@storybook/manager-webpack5": "6.5.9",
+    "@storybook/react": "6.5.9",
     "@teamsupercell/typings-for-css-modules-loader": "2.5.1",
     "@types/bootstrap": "4.6.0",
     "@types/case-sensitive-paths-webpack-plugin": "2.1.6",
@@ -116,50 +99,36 @@
     "@types/debounce": "1.2.1",
     "@types/diff-match-patch": "1.0.32",
     "@types/fs-extra": "9.0.13",
-    "@types/jest": "27.0.3",
-    "@types/jquery": "3.5.10",
-    "@types/katex": "0.11.1",
-    "@types/mini-css-extract-plugin": "2.4.0",
-    "@types/react": "17.0.37",
+    "@types/jest": "28.1.2",
+    "@types/jquery": "3.5.14",
+    "@types/katex": "0.14.0",
+    "@types/mini-css-extract-plugin": "2.5.1",
+    "@types/react": "18.0.14",
     "@types/react-cookies": "0.1.0",
-    "@types/react-dev-utils": "9.0.9",
-    "@types/react-dom": "17.0.11",
-    "@types/react-helmet": "6.1.4",
+    "@types/react-dev-utils": "9.0.11",
+    "@types/react-dom": "18.0.5",
+    "@types/react-helmet": "6.1.5",
     "@types/react-image-gallery": "1.0.5",
-    "@types/react-redux": "7.1.20",
-    "@types/react-responsive": "8.0.4",
-    "@types/react-router": "4.4.5",
-    "@types/react-router-dom": "5.3.1",
+    "@types/react-responsive": "8.0.5",
     "@types/react-transition-group": "4.4.4",
     "@types/react-youtube": "7.10.0",
     "@types/redux-mock-store": "1.0.3",
+    "@types/webpack-bundle-analyzer": "4.4.1",
     "@types/workbox-webpack-plugin": "5.1.8",
     "@types/youtube": "0.0.46",
-    "@typescript-eslint/eslint-plugin": "4.33.0",
-    "@typescript-eslint/parser": "4.33.0",
-    "creevey": "0.7.30",
-    "eslint": "7.32.0",
-    "eslint-config-react-app": "6.0.0",
-    "eslint-loader": "4.0.2",
-    "eslint-plugin-flowtype": "6.1.1",
-    "eslint-plugin-import": "2.25.2",
-    "eslint-plugin-jsx-a11y": "6.4.1",
-    "eslint-plugin-react": "7.26.1",
-    "eslint-plugin-react-hooks": "4.2.0",
-    "jest": "27.4.3",
-    "loki": "0.28.1",
+    "@typescript-eslint/eslint-plugin": "5.29.0",
+    "@typescript-eslint/parser": "5.29.0",
+    "creevey": "0.7.39",
+    "eslint": "8.18.0",
+    "eslint-plugin-storybook": "0.5.12",
+    "jest": "28.1.1",
+    "jest-environment-jsdom": "28.1.1",
+    "loki": "0.30.3",
     "redux-mock-store": "1.5.4",
-    "ts-jest": "27.1.1"
-  },
-  "eslintConfig": {
-    "extends": "react-app",
-    "plugins": [
-      "react-hooks"
-    ],
-    "rules": {
-      "react-hooks/rules-of-hooks": "error",
-      "react-hooks/exhaustive-deps": "warn"
-    }
+    "ts-jest": "28.0.5",
+    "ts-node": "10.8.1",
+    "webpack-cli": "4.10.0",
+    "webpack-dev-server": "4.9.2"
   },
   "jest": {
     "preset": "ts-jest",
