@@ -1,7 +1,8 @@
-const path = require('path');
-const paths = require('./paths');
+import path from "path";
+import { Configuration } from "webpack";
+import paths from "./paths";
 
-module.exports = {
+export default {
 	resolve: {
 		alias: {
 			ui: path.resolve(paths.appNodeModules, '@skbkontur/react-ui'),
@@ -10,4 +11,4 @@ module.exports = {
 			//if u adding a new one don't forget to add it to package.json jest.moduleNameMapper section, ot *.test.* won't work
 		}
 	},
-};
+} as Configuration;
