@@ -1,9 +1,10 @@
 import React from 'react';
-import ReactDOM from 'react-dom';
 import UlearnApp from "./App";
+import { createRoot } from 'react-dom/client';
 
 it('renders without crashing', () => {
-	const div = document.createElement('div');
-	ReactDOM.render(<UlearnApp />, div);
-	ReactDOM.unmountComponentAtNode(div);
+	const container = document.createElement('div');
+	const root = createRoot(container);
+	root.render(<UlearnApp/>);
+	root.unmount();
 });
