@@ -1,4 +1,4 @@
-import config from "src/config";
+import config from "src/proxyConfig";
 
 import { Toast } from "ui";
 
