@@ -1,5 +1,5 @@
-import "katex/dist/katex.min.css";
 import renderMathInElement from 'katex/dist/contrib/auto-render.js';
+import "katex/dist/katex.min.css";
 
 export default function translateTextToMathKatex(element: HTMLElement, additionalSettings: Record<string, unknown>): void {
 	const text = element.innerText;
