@@ -1,5 +1,5 @@
 import React, { Component } from "react";
-import moment from "moment";
+import moment from "moment-timezone";
 
 import { Hint, } from "ui";
 import { Link } from "react-router-dom";
