@@ -1,10 +1,13 @@
 import React from "react";
+import { ThemeContext } from "ui";
+import { PC, } from "icons";
 
 import ShowControlsTextContext from "./ShowControlsTextContext";
 import IControlWithText from "./IControlWithText";
 
-import { PC, } from "@skbkontur/react-icons";
 import { Visualizer } from "../Visualizer/Visualizer";
+
+import defaultTheme from "src/uiTheme";
 
 import styles from './Controls.less';
 
@@ -24,20 +27,22 @@ function VisualizerButton({
 	const [isModalVisible, setModalVisible] = React.useState(false);
 
 	return (
-		<span className={ styles.exerciseControls }>
-			<ShowControlsTextContext.Consumer>
-			{
-				(showControlsTextContext) =>
-					<span onClick={ openModal }>
-						<span className={ styles.exerciseControlsIcon }>
-							<PC/>
+		<ThemeContext.Provider value={ defaultTheme }>
+			<span className={ styles.exerciseControls }>
+				<ShowControlsTextContext.Consumer>
+				{
+					(showControlsTextContext) =>
+						<span onClick={ openModal }>
+							<span className={ styles.exerciseControlsIcon }>
+								<PC/>
+							</span>
+							{ (showControlsTextContext || showControlsText) && texts.controls.visualizer.text }
 						</span>
-						{ (showControlsTextContext || showControlsText) && texts.controls.visualizer.text }
-					</span>
-			}
-			</ShowControlsTextContext.Consumer>
-			{ isModalVisible && <Visualizer code={ code } onModalClose={ closeModal }/> }
-		</span>
+				}
+				</ShowControlsTextContext.Consumer>
+				{ isModalVisible && <Visualizer code={ code } onModalClose={ closeModal }/> }
+			</span>
+		</ThemeContext.Provider>
 	);
 
 	function openModal() {
