@@ -14,9 +14,13 @@ const defaultProps: Props = {
 	id: '1',
 	isActive: false,
 	title: "Первое знакомство с C#",
+	courseId: 'basicprogramming',
+	isNotPublished: true,
+	publicationDate: undefined,
+	additionalContentInfo: { isAdditionalContent: false, publicationDate: null, },
 	progress: {
-		statusesBySlides: {},
 		current: 0,
+		additionalInfoBySlide: {},
 		max: 100,
 		inProgress: 0,
 	},
@@ -41,8 +45,9 @@ Default.args = [
 	{ title: 'Текущий модуль', isActive: true, },
 	{ title: "Модуль с прогресс-баром", },
 	{
-		title: "Модуль с заполненным прогресс-баром", progress: {
-			statusesBySlides: {},
+		title: "Модуль с заполненным прогресс-баром",
+		progress: {
+			additionalInfoBySlide: {},
 			current: 100,
 			max: 100,
 			inProgress: 0,
