@@ -1,6 +1,8 @@
+import isInDevelopment from "../isInDevelopment";
+
 let tz = 'Europe/Moscow';
 
-if(process.env.NODE_ENV === 'development') {
+if(isInDevelopment) {
 	tz = Intl.DateTimeFormat().resolvedOptions().timeZone;
 }
 export const DEFAULT_TIMEZONE = tz;
