@@ -1,4 +1,4 @@
-import { FLAT_THEME_OLD, ThemeFactory, } from "ui";
+import { DEFAULT_THEME, ThemeFactory, } from "ui";
 
 const roundButtons = {
 	btnBorderRadiusLarge: '8px',
@@ -29,14 +29,14 @@ const reviewReplyTextarea = {
 export default ThemeFactory.create({
 	...roundButtons,
 	...roundSwitcher
-}, FLAT_THEME_OLD);
+}, DEFAULT_THEME);
 
 export const textareaHidden = ThemeFactory.create({
 	...reviewReplyTextarea,
-}, FLAT_THEME_OLD);
+}, DEFAULT_THEME);
 
 //currently it only applies a dark background and white text in popup(tooltip), styles copied from dark theme
 export const darkFlat = ThemeFactory.create({
 	...darkPopups,
 	...roundButtons,
-}, FLAT_THEME_OLD);
+}, DEFAULT_THEME);
